anti dupe reward stuff (but it works properly now)

+++ code/WorkInProgress/dialogue/dialogue_cultist_base.dm
@@ -416,9 +416,12 @@
 			else return 0
 
 		onActivate(var/client/C)
+			if(QDELETED(src))
+				return
 			var/obj/dialogueobj/controlpc/M = new /obj/dialogueobj/controlpc
 			M.lever_lv(C.mob)
-			qdel(src) // pretty sure all of these are needed for cleanup
+			for(var/datum/dialogueNode/nodes in src.master.allNodes)
+				qdel(nodes)
 			qdel(src.master)
 			qdel(src.master.master)
 
@@ -433,9 +436,12 @@
 			else return 0
 
 		onActivate(var/client/C)
+			if(QDELETED(src))
+				return
 			var/obj/dialogueobj/controlpc/M = new /obj/dialogueobj/controlpc
 			M.lever_hv(C.mob)
-			qdel(src) // pretty sure all of these are needed for cleanup
+			for(var/datum/dialogueNode/nodes in src.master.allNodes)
+				qdel(nodes)
 			qdel(src.master)
 			qdel(src.master.master)
